CPLP-1419 cleanup
