+++ src/marketplace/CatenaX.NetworkServices.App.Service/ViewModels/AppEditableDetail.cs
@@ -1,4 +1,4 @@
-﻿/********************************************************************************
+/********************************************************************************
  * Copyright (c) 2021,2022 BMW Group AG
  * Copyright (c) 2021,2022 Contributors to the CatenaX (ng) GitHub Organisation.
  *
@@ -18,30 +18,23 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-namespace CatenaX.NetworkServices.PortalBackend.DBAccess.Models;
+namespace CatenaX.NetworkServices.App.Service.ViewModels;
 
 /// <summary>
-///  View model of an application's base data.
+/// Model for updating an app.
 /// </summary>
-public record InReviewAppData(string Title, string Provider, string LeadPictureUri)
-{
+/// <param name="Descriptions"></param>
+/// <param name="Images"></param>
+/// <param name="ProviderUri"></param>
+/// <param name="ContactEmail"></param>
+/// <param name="ContactNumber"></param>
+/// <returns></returns>
+public record AppEditableDetail(IEnumerable<Localization> Descriptions, IEnumerable<string> Images, string? ProviderUri, string? ContactEmail, string? ContactNumber);
 
-    /// <summary>
-    /// ID of the app.
-    /// </summary>
-    public Guid Id { get; set; }
-
-    /// <summary>
-    /// Title or name of the app.
-    /// </summary>
-    public string Title { get; set; } = Title;
-    /// <summary>
-    /// Provider of the app.
-    /// </summary>
-    public string Provider { get; set; } = Provider;
-
-    /// <summary>
-    /// Uri to app's lead picture.
-    /// </summary>
-    public string LeadPictureUri { get; set; } = LeadPictureUri;
-}
+/// <summary>
+/// Model for LanguageCode and Description
+/// </summary>
+/// <param name="LanguageCode"></param>
+/// <param name="LongDescription"></param>
+/// <returns></returns>
+public record Localization(string LanguageCode, string LongDescription);
